qcom-rradc: Initialize USBIN_V and BATT_ID in kthread
Either it's a recent change in Ubuntu Touch or faulty hardware,
but bootup is slowed down dramatically now due to the RRADC not
answering channel reads anymore.

To improve this situation, try to read and cache the values
in question in a kthread and save them for later use.

This should also help subsequent battery stat reads.